requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.3
yarl==1.7.2
